pyproject.toml
[project]
name = "build-agent"
version = "0.1.0"
description = "A general-purpose AI Agent following 12-step progression"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]
